refactoring

+++ src/main/java/it/polimi/parkingService/webApplication/parking/dao/ParkingAreaRepository.java
@@ -3,7 +3,7 @@ package it.polimi.parkingService.webApplication.parking.dao;
 import it.polimi.parkingService.webApplication.parking.models.ParkingArea;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParkingAreaRepository extends JpaRepository<ParkingArea, Integer> {
+public interface ParkingAreaRepository extends JpaRepository<ParkingArea, Long> {
 
     ParkingArea findByName(String name);
 }
